django 200625

+++ Django/formclass/articles/views.py
@@ -1,6 +1,7 @@
 from django.shortcuts import render, redirect
 from .models import Article, Comment
 from .forms import ArticleForm, CommentForm
+from django.contrib.auth.decorators import login_required
 
 # Create your views here.
 def index(request):
@@ -10,18 +11,26 @@ def index(request):
     }
     return render(request, 'articles/index.html', context)
 
+# 로그인 검증
+# o -> new 함수 실행
+# x -> login 페이지로 보냄
+@login_required 
 def new(request):
     if request.method == 'POST':
         # Database에 저장
         # 1. 요청에 실려온 data 꺼내오기
         # title = request.POST.get('title')
         # content = request.POST.get('content')
-        form = ArticleForm(request.POST)
+        # image = request.FILES.get('image')
+        form = ArticleForm(request.POST, request.FILES)
         
         # 2-1. data 유효성 검사
         if form.is_valid():
             # (ModelForm) 2-2. Database에 저장
-            article = form.save()
+            article = form.save(commit=False)
+            article.user = request.user
+            # article.user_id = request.uesr.pk
+            article.save()
             # # 2-2.검증된 data 꺼내오기
             # title = form.cleaned_data.get('title')
             # content = form.cleaned_data.get('content')
@@ -52,15 +61,24 @@ def detail(request, pk):
     }
     return render(request, 'articles/detail.html', context)
 
+@login_required 
 def delete(request, pk): # POST
     article = Article.objects.get(pk=pk)
+
+    if request.user != article.user:
+        return redirect('articles:detail', article.pk)
+
     if request.method == 'POST':
         article.delete()
     return redirect('articles:index')
 
+@login_required 
 def edit(request, pk):
     # 1. Database에서 data 가져오기
     article = Article.objects.get(pk=pk)
+
+    if request.user != article.user:
+        return redirect('articles:detail', article.pk)
 
     if request.method == 'POST':
         # data 수정!
@@ -93,6 +111,7 @@ def edit(request, pk):
     }
     return render(request, 'articles/edit.html', context)
 
+@login_required
 def comments_new(request, article_pk):
     # 1. 요청이 POST인지 점검
     if request.method == 'POST':
@@ -105,23 +124,35 @@ def comments_new(request, article_pk):
             comment = form.save(commit=False)
             # 4-1. article 정보 주입
             comment.article_id = article_pk
+            # 4-2. user 정보넣기
+            comment.user=request.user
             comment.save()
     # 5. 생성된 댓글을 확인할 수 있는 곳으로 안내
     return redirect('articles:detail', article_pk)
 
+@login_required
 def comments_delete(request, article_pk, pk): # 'POST'
     # 0. 요청이 POST인지 점검
+    # 1. pk를 가지고 삭제하려는 data 꺼내오기
+    comment = Comment.objects.get(pk=pk)
+
+    if request.user != article.user:
+        return redirect('articles:detail', comment.article.pk)
+
     if request.method == 'POST':
-        # 1. pk를 가지고 삭제하려는 data 꺼내오기
-        comment = Comment.objects.get(pk=pk)
         # 2. 삭제
         comment.delete()
     # 3. 삭제되었는지 확인 가능한 곳으로 안내
     return redirect('articles:detail', article_pk)
 
+@login_required
 def comments_edit(request, article_pk, pk): # GET, POST
     # Database에서 수정하려 하는 data 가져오기
     comment = Comment.objects.get(pk=pk)
+
+    if request.user != article.user:
+        return redirect('articles:detail', comment.article.pk)
+
     # 0. 요청의 종류가 POST인지 GET인지 점검
     if request.method == 'POST':
         # 실제로 수정!
@@ -142,3 +173,16 @@ def comments_edit(request, article_pk, pk): # GET, POST
         'form': form,
     }
     return render(request, 'articles/comments_edit.html', context)
+
+def like(request, pk):
+    user = request.user
+    article = Article.objects.get(pk=pk)
+
+    if request.method == 'POST':
+        if user in article.like_users.all():
+            article.like_users.remove(user)
+        else:
+            article.like_users.add(user)
+
+
+    return redirect('articles:detail', pk)
